archive/notebooklm-automation/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notebooklm-automation"
version = "1.0.0"
description = "Unified NotebookLM automation — bridge, health, adapters, publishers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "biojuho"}]

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "loguru>=0.7",
    "pydantic>=2.5",
]

[project.optional-dependencies]
notebooklm = [
    "notebooklm-py>=0.1.0",
]
extraction = [
    "PyMuPDF>=1.24",
    "pdfplumber>=0.11",
]
ocr = [
    "google-cloud-vision>=3.5",
    "pytesseract>=0.3",
    "Pillow>=10.0",
]
slides = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "python-pptx>=0.6",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "respx>=0.21",
]

[project.scripts]
notebooklm-health = "notebooklm_automation.health:_cli_main"
notebooklm-api = "notebooklm_automation.api:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring real NotebookLM API",
    "external: marks tests requiring external services",
]
